Data Structure solution 1
Implement a function that accepts two integers arrays 'arr1' and 'arr2' of sizes n and m respectively as its argument to find and return the sum of all uncommon elements in two arrays(elements which are present in only one of the array).Print -1 if both arrays are null. If one of the arrays is null then return the sum of all elements of the other array.